pyproject.toml
[project]
name = "mujibswap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "onnxruntime>=1.22.1",
    "python-multipart>=0.0.20",
    "socaity-face2face>=1.1.8",
]
